pyproject.toml
[project]
name = "openenv-scalar"
version = "0.1.0"
description = "OpenEnv Email Triage challenge implementation"
dependencies = [
    "fastapi",
    "uvicorn",
    "openai",
    "openenv-core",
    "pydantic",
    "python-dotenv",
]

[project.scripts]
server = "server.app:main"
inference = "inference:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
